flask-demo/requirements.txt
Flask
Jinja2
pytest
pyyaml
markdown
google-cloud-aiplatform
